requirements.txt
aiohttp==1.3.5
amqp==2.2.2
appdirs==1.4.3
appnope==0.1.0
async-timeout==3.0.0
attrs==18.1.0
Babel==2.5.3
billiard==3.5.0.3
bingads==12.0.3.1
cached-property==1.4.3
cachetools==3.1.1
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
colorclass==2.2.0
coverage==4.5.2
decorator==4.1.2
defusedxml==0.5.0
Django==2.1.1
django-admin-honeypot==1.1.0
django-celery-beat==1.3.0
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-crontab==0.7.1
django-extensions==1.9.7
django-timezone-field==3.0
djangorestframework==3.7.3
djangorestframework-simplejwt==3.2.3
docopt==0.6.2
enum34==1.1.6
facebook-business==3.0.0
facebookads==2.11.4
flower==0.9.3
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
googleads==15.0.2
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
ipython-genutils==0.2.0
isodate==0.6.0
kombu==4.2.1
lxml==4.2.3
multidict==4.4.2
oauth2client==4.1.2
oauthlib==2.1.0
packaging==19.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pip-upgrader==1.4.6
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
Pygments==2.2.0
PyJWT==1.6.4
pyparsing==2.3.1
PySocks==1.6.7
python-crontab==2.3.6
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==4.2b4
pyzmq==17.1.2
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==4.0
simplegeneric==0.8.1
six==1.12.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
suds-jurko==0.6
terminaltables==3.1.0
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.12.2
xmltodict==0.11.0
yarl==0.9.8
zeep==3.0.0
